Renamed fragments to templates
