+++ client/src/components/templates/__tests__/SocialSignInButton.test.tsx
